Resolve precedent bug introduced
introduced a new implementation type DVarInt, and the all test is green, I should be happy but instead I'm wondering how all this happened

+++ structure/type/DVarInt.h
@@ -6,16 +6,19 @@
 #define PARSINGBLOCKCHAIN_CVARINT_H
 
 #include <fstream>
+#include "../../util/serialize.h"
 
 class DVarInt{
 
 private:
     uint64_t value;
+    //CCompactSize newValue;
 
 public:
+
     virtual ~DVarInt();
 
-    int getValue() const;
+    uint64_t getValue() const;
 
     void decode(std::ifstream &stream);
 };
